Move CommandKey related classes to its own command_key module
